refactor: Moved to controller-service-repo based architecture. Added common project to store dto models.
