+++ Bnb.Services/Bnb.Services.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net9.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Bnb.Common\Bnb.Common.csproj" />
+      <ProjectReference Include="..\Bnb.Entities\Bnb.Entities.csproj" />
+      <ProjectReference Include="..\Bnb.Repos\Bnb.Repos.csproj" />
+    </ItemGroup>
+
+</Project>
