+++ Bnb.Services/IUserService.cs
@@ -1,0 +1,9 @@
+using Bnb.Common.Dtos.Responses;
+
+namespace Bnb.Services;
+
+public interface IUserService
+{
+    Task<IEnumerable<UserDto>> GetUsersAsync();
+    Task<UserDto?> GetUserByIdAsync(int id);
+}
